+++ src/main/java/com/almostreliable/summoningrituals/recipe/condition/BlockStateCondition.java
@@ -10,11 +10,8 @@ import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.world.level.storage.loot.predicates.LootItemBlockStatePropertyCondition;
 
-import org.jetbrains.annotations.Nullable;
-
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 public class BlockStateCondition implements ConditionHandler<LootItemBlockStatePropertyCondition> {
 
@@ -25,7 +22,7 @@ public class BlockStateCondition implements ConditionHandler<LootItemBlockStateP
         LootItemBlockStatePropertyCondition::new
     );
 
-    private static final LangEntry BLOCK_STATE = LangEntry.condition("block_state", "Block Properties");
+    private static final LangEntry ALTAR_PROPERTIES = LangEntry.condition("altar_properties", "Altar Properties");
 
     @Override
     public StreamCodec<RegistryFriendlyByteBuf, LootItemBlockStatePropertyCondition> getStreamCodec() {
@@ -40,45 +37,10 @@ public class BlockStateCondition implements ConditionHandler<LootItemBlockStateP
         var propertyTooltips = new ArrayList<Component>();
         var propertyPredicate = opt.get();
         var properties = propertyPredicate.properties();
-
-        for (var prop : properties) {
-            var propName = Component.literal(prop.name());
-            var matcher = prop.valueMatcher();
-
-            if (matcher instanceof StatePropertiesPredicate.ExactMatcher(var exactValue)) {
-                propertyTooltips.add(conditionNamedValueComponent(propName, exactValue));
-                continue;
-            }
-
-            if (matcher instanceof StatePropertiesPredicate.RangedMatcher(var minValueOpt, var maxValueOpt)) {
-                var value = getRangedValueString(minValueOpt, maxValueOpt);
-                if (value == null) continue;
-
-                propertyTooltips.add(conditionNamedValueComponent(propName, value));
-            }
-        }
+        addBlockStateTooltip(propertyTooltips, properties);
 
         if (propertyTooltips.isEmpty()) return;
-        tooltip.add(conditionNameComponent(BLOCK_STATE.get()));
+        tooltip.add(conditionNameComponent(ALTAR_PROPERTIES.get()));
         tooltip.addAll(propertyTooltips);
-    }
-
-    @SuppressWarnings({"OptionalUsedAsFieldOrParameterType", "OptionalIsPresent"})
-    @Nullable
-    private static String getRangedValueString(Optional<String> minValueOpt, Optional<String> maxValueOpt) {
-        if (minValueOpt.isPresent() && maxValueOpt.isPresent()) {
-            var minValue = minValueOpt.get();
-            var maxValue = maxValueOpt.get();
-            if (minValue.equals(maxValue)) {
-                return minValue;
-            }
-            return minValue + " - " + maxValue;
-        }
-
-        if (minValueOpt.isPresent()) {
-            return minValueOpt.get();
-        }
-
-        return maxValueOpt.orElse(null);
     }
 }